+++ commitlint.config.js
@@ -1,6 +1,18 @@
+const scopes = require('@commitlint/config-lerna-scopes')
+
+const prefix = 'ps-design-system-'
+
 module.exports = {
-  extends: ['@commitlint/config-angular', '@commitlint/config-lerna-scopes'],
+  extends: ['@commitlint/config-angular'],
   rules: {
-    lang: [0]
+    lang: [0],
+    'scope-enum': ctx =>
+      scopes.rules['scope-enum'](ctx).then(([level, applicable, packages]) => [
+        level,
+        applicable,
+        packages.map(
+          pkg => (pkg.includes(prefix) ? pkg.replace(prefix, '') : pkg)
+        )
+      ])
   }
 }