+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@1.5.3...@pluralsight/ps-design-system-drawer@1.5.4) (2018-10-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
 <a name="1.5.3"></a>
 ## [1.5.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@1.5.2...@pluralsight/ps-design-system-drawer@1.5.3) (2018-10-31)