+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-radio@0.3.7...@pluralsight/ps-design-system-radio@0.3.8) (2018-10-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-radio
+
+
+
+
+
 <a name="0.3.7"></a>
 ## [0.3.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-radio@0.3.6...@pluralsight/ps-design-system-radio@0.3.7) (2018-10-31)